Refactor: apply useOnClickOutside hook to dropdowns in `MoreActionsMenu.tsx`, `ModelSelect.tsx`, and `VerbositySelect.tsx`
