+++ src/renderer/components/MoreActionsMenu.tsx
@@ -3,6 +3,7 @@ import classNames from "classnames";
 import React, { useEffect, useRef, useState } from "react";
 import { Link, useNavigate } from "react-router-dom";
 import { Tooltip } from "react-tooltip";
+import { useOnClickOutside } from "../helpers";
 import { useAppDispatch, useAppSelector } from "../hooks";
 import { useMessenger } from "../send-to-backend";
 import { RootState } from "../store";
@@ -35,28 +36,11 @@ export default function MoreActionsMenu({
   const backendMessenger = useMessenger(vscode);
   const [showViewOptions, setShowViewOptions] = useState(false);
 
-  // Reference to the menu container for outside click detection
-  const menuRef = useRef<HTMLDivElement>(null);
+  // Reference to the dropdown container for outside click detection
+  const dropdownRef = useRef<HTMLDivElement>(null);
 
-  // Close the actions menu when clicking outside of it
-  useEffect(() => {
-    function handleClickOutside(event: MouseEvent) {
-      if (menuRef.current && !menuRef.current.contains(event.target as Node)) {
-        setShowMoreActions(false);
-      }
-    }
-
-    if (showMoreActions) {
-      document.addEventListener("mousedown", handleClickOutside);
-    } else {
-      document.removeEventListener("mousedown", handleClickOutside);
-    }
-
-    return () => {
-      document.removeEventListener("mousedown", handleClickOutside);
-    };
-  }, [showMoreActions]);
-
+  // Close the dropdown menu when clicking outside of it
+  useOnClickOutside(dropdownRef, () => setShowMoreActions(false));
 
   // When the show more actions menu is shown, hide the view options menu
   // So, if the user leaves the view options menu open on close, it won't be shown on open
@@ -70,7 +54,7 @@ export default function MoreActionsMenu({
     <>
       <div
         id="more-actions-menu"
-        ref={menuRef}
+        ref={dropdownRef}
         className={classNames(
           "MoreActionsMenu",
           "fixed z-20 right-4 p-2 bg-menu rounded border border-menu overflow-hidden max-w-[calc(100vw-2em)]",
